fix: escape External Secrets template syntax in github-pat-externalsecret
When moved to Helm templates directory, the External Secrets template syntax
'{{ .token }}' was being evaluated by Helm as a template variable, resulting
in an empty password field.

Escaped the syntax using Helm's template escaping: '{{ "{{" }} .token {{ "}}" }}'
so it's treated as a literal string and passed through to the ExternalSecret
resource where it will be correctly evaluated by External Secrets Operator.

This fixes ArgoCD's ability to access the segment-deployment repository.